Configuration, REST definition, protocol picker
Added a resource file that specifies basic configuration information such
as the Thredds server URL or the name of different protocols within the
resulting URI.

Added a utility class that is able to read properties from the
configuration file (ressources/config.properties)

Added a setup and destruction hook to initiate and destroy resources (the
configuration reader)

Added the dataset in the request URL.

Parsing of query parameters: simply concatenate all query parameters to
obtain a single query which is semantically parsed in the next step.

Added the protocol picker class which ought to map a request to any other
protocol. Currently it simply maps to OPeNDAP only without any logic
involved.