Variable length dataset specifier, first mapping
Since datasets are identified as a virtual URI on the
Thredds catalogue the URL needs to specify a flexible length
path parameter. Created such a path and concatenates all parts of the URI
before processing the request.

The rest service can now be reached directly as ../EasyThredds/...
So the 'rest' part was dropped.

Created the first implementation of a mapping for OpENDAP.