Added custom variables, started NCSS
Started to implement the mapping to the NCSS protocol.
Therefore, a bounding box is created and the required variables
are mentioned.

Added a variable declaration to the custom format: var=a,b,c
will extract the variables a, b and c from the dataset.

The protocol picker will now cycle through all implemented protocols
returning the translation one at a time.