Implemented CmdRemote, corrected NCSS & OPeNDAP
Added a mapping to the CmdRemote protocol.

Extended the mapping to the ncss protocol to provide the hight level
through the vertical stride or coordinate.

Corrected the OPeNDAP protocol to provide the 4 dimensions for each
variable if present.

Facilitated the range definitions. If no end point is defined the start
point is used as an end. Thereby, single value ranges can be defined.
For example, [a::] specifies a range containing only the point a.