File extensions, time / space ranges
Added data file extension:
Each protocol can specify an extension which is attachced to the data set
directory. E.g. OPeNDAP can extend .ascii to get ascii feedback.

NetCDF output for NCSS:
added a new property which can set the output format to be returned by the
ncss request. netcdf is predefined but can be changed to netcdf4.

Extended the range specifications of the custom protocol:
Time ranges are now specified as a start time, an end time, a start index,
an end index and a stride. The start and end time define the interval. The
start and end index define the first and last record in this interval. The
stride determines at which stride the remaining interval is traversed.

Similary a spatial range was introduced to handle definitions of the
latitude and longitude together with an interval selection.

Those two types are necessary as some protocols require index intervals
and others require spatial / time intervals. A combination is possible.