Implemented OPeNDAP file reader
In order to measure the performance of the OPeNDAP protocol the query is
executed and a .dods file is obtained for download. Next, the .dds file is
acquired. With both of these files available the data can be iterated
locally.

Neither .dods, nor .dap, nor .dds files are commited to version control.

Added functionality to count the number of bytes iterated.