Implemented CdmRemote reading using generics
Implemented the CdmRemote reading ability:
- requesting the ncstream using the factory method
- reading the ncstream using the NcStreamReader
- unfortunately, the return type with the gathered data is not accessible
  so need to use generics to access the data

This was the only working way to obtain the result stream:
1. using NetcdfDataset.openDataset("cdmremote:...") appends req=header as
a request. Further triggered methods fails as there is no data in the
response. This is required by the internal setup of the CdmRemote file. A
MAGIC_DATA header is missing!
2. Directly loading the stream is incorrect as several header fields are
included. The parsing is done in the NcStreamReader

Changed to version 4.8.0 for all netcdf related artefacts.
Changed the default output format of CmdRemote to ncstream.

Corrected the byte size:
- it depends on the output format of the data which can be obtained
  dynamically