Time translation, fixed OPeNDAP requests
Had to split the meta data download into two parts for OPeNDAP:
- first lat,lon,lev are fetched and placed into a dods file
- next, only time is queried and place into a dods file
- while reading the index series different readers can be used
-> this process was necessary as the NetCdf library was confused with the
returned data types of the different formats.

Implemented a translation from the weird time format provided as days
since 1950-01-01 00:00:00 as a floating point number (where 0.5 is half a
day).

When requesting a netcdf file an identifier can be provided to store a
unique file.

Improved error output:
- when absolute values are out of bounds the resolved index range is
  displayed
- returning full error reports instead of an error message only to make
  debugging easier