Extended NCSS, test cases, canTranslate()
All ranges can now indicate whether they are a point.

All translated protocols have the possibility to state whether a given
collective protocol can be translated or not:
- NCSS does not provide different strides on latitude and longitude
- NCSS can only do subsetting on the altitude of the entire range (min
  value to max value) is used

Added special cases for NCSS where the ranges are points.
Added a flag that ensures CF conformity for netcdf output.

Added utility methods to the NetCdfReader that provide size / length
information on the variables.

Created a test class that ensures that the required variables are present
and that they have the same cardinality and extent.

Added a test utility that removes the auxiliary files from disk after each
unit test.