NCSS meta data loading, single point checks
Added different datasets and variables for [1-4]D random query generation.

Moved the fetching of the VariableReader into the super class and the base
class is only called if the dimension data actually needs to be fetched.

Implemented dimension loading for NCSS:
- ncss does not support queries on dimension variables
- instead if provides meta data by appending /dataset.xml to the dataset
- from this XML-file the axis data can be collected using a custom parser
- had to add a meta reader that is capable of parsing the inputs for NCSS

Added a utility method that can check if a given absolute range only
comprises a single index point.

Better output for performance measurement using a console progress bar
(library)